[UPD] Moved translate context and global provider to own folder
