+++ app/components/module/main/header/MainMenu/MainMenu.tsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 
 import Menu from 'antd/lib/menu';
-import { TranslateContext } from 'modules/TranslateContext';
+import { TranslateContext } from 'modules/global/TranslateContext';
 import IOMSLDMainMenu from 'data/common/model/lang/main_window/OMSLDMainMenu.interface';
 
 export default class MainMenu extends PureComponent {
