+++ app/modules/main/content/AppStart/AppStart.tsx
@@ -5,7 +5,7 @@ import Col from 'antd/lib/col';
 import Typography from 'antd/lib/typography';
 
 import styles from './AppStart.m.scss';
-import { TranslateContext } from 'modules/TranslateContext';
+import { TranslateContext } from 'modules/global/TranslateContext';
 import IOMSLDMainStart from 'data/common/model/lang/main_window/OMSLDMainStart.interface';
 import StartFrameActionLink from 'components/module/main/content/StartFrameActionLink/StartFrameActionLink';
 import { EOmsIconIconName } from 'components/common/OMSIcon/OMSIcon.icon-enum';
