+++ app/modules/main/modals/AboutAppModal/AboutAppModal.tsx
@@ -4,7 +4,7 @@ import Modal from 'antd/lib/modal';
 import Typography from 'antd/lib/typography';
 
 import { oms } from 'data/data.init';
-import { TranslateContext } from 'modules/TranslateContext';
+import { TranslateContext } from 'modules/global/TranslateContext';
 import IOMSLDAboutAppModal from 'data/common/model/lang/main_window/OMSLDAboutAppModal.interface';
 import { APP_CONSTS } from 'data/utils/AppConsts.enum';
